+++ js/IndexDirective.js
@@ -5,9 +5,6 @@
   /**
    * Creates a new Angular directive as an HTML element ov-index to create an openVeo player
    * index, with a list of presentation slides.
-   * It requires ovPlayerDirectory global variable to be defined and have
-   * a value corresponding to the path of the openVeo Player
-   * root directory.
    *
    * e.g.
    * <ov-index></ov-index>
@@ -19,7 +16,7 @@
     return {
       require: ['^ovPlayer', '^ovPlayerTabs'],
       restrict: 'E',
-      templateUrl: ovPlayerDirectory + 'templates/index.html',
+      templateUrl: 'ov-player-index.html',
       scope: true,
       link: ovIndexLink
     };