+++ tasks/player.js
@@ -2,6 +2,7 @@
 
 module.exports = {
   basePath: ['.'],
+  templatesPath: ['<%= player.basePath %>/templates/'],
   jsPath: ['<%= player.basePath %>/js/'],
   distPath: ['<%= player.basePath %>/dist'],
   sass: ['<%= player.basePath %>/css']